% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MQMF.r
\docType{data}
\name{abdat}
\alias{abdat}
\title{abdat Three data objects suitable for use with MQMF.}
\format{A list of three objects
\describe{
  \item{spsname}{ the name of the species concerned}
  \item{resilience}{ a proxy for the relative productivity, for
        use when using catch-MSY }
  \item{fish}{ a data.frame containing year, catch, and cpue}
}}
\usage{
abdat
}
\description{
A dataset containing the fish data.frame, the
    resilience, and the spsname set up ready for use with MQMF.
    In particular it can be used when fitting a surplus
    production model. Workable initial parameter values, before log-
    transformation might be: r= 0.42,K=9400,Binit=3400,sigma=0.05
}
\keyword{datasets}
