% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm.r
\name{altnegLL}
\alias{altnegLL}
\title{altnegLL calculate the normal negative log-likelihood}
\usage{
altnegLL(inp, indat)
}
\arguments{
\item{inp}{a vector of model parameters (r,K,Binit)}

\item{indat}{a matrix with at least columns 'year', 'catch', and 'cpue'}
}
\value{
a single value the negative log-likelihood
}
\description{
altnegLL calculates negLLM using the simplification
   from Haddon (2011) using the ssq calculated within the function
   spm
}
\examples{
\dontrun{
data(dataspm)
fish <- dataspm$fish
head(fish,15)
colnames(fish) <- tolower(colnames(fish))
pars <- c(r=0.242,K=5170,Binit=2840)
altnegLL(pars,fish)  # should be -12.1249
}
}
