% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MQMF.r
\docType{data}
\name{blackisland}
\alias{blackisland}
\title{blackisland tagging data from an abalone population}
\format{A data.frame of six columns
\describe{
  \item{deltat}{the time between tagging and recapture, in years}
  \item{time1}{the date at tagging}
  \item{len1}{the shell length when tagged in mm}
  \item{time2}{the date of recapture}
  \item{len2}{the shell length at recapture in mm}
  \item{deltal}{the growth increment between tagging and recapture
                in mm; there are a few negative values.}
}}
\usage{
blackisland
}
\description{
A dataset containing the time1 of tagging, the time2
    of tag return (approximately 1 year between them), the shell 
    length at tagging, len1, and at recapture, len2, with the time
    interval, deltat, and the growth increment, deltal. This data
    can be used to estimate the growth characteristics of abalone
    from the Black Island site, which is on the south west coast of
    Tasmania, Australia. Thanks to Dr Craig Mundy and the abalone 
    team at the Institute of Marine and Antarctic Studies, of the
    University of Tasmania for the use of this data. The time 
    interval between tagging and recapture is 1 year and 1 week,
    1.02 years, which reflects the practical problems of taking a 
    vessel around the bottom of Tasmania, where it is essential to
    wait on suitable weather for such sub-tidal field work.
}
\keyword{datasets}
