% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{calcprior}
\alias{calcprior}
\title{calcprior return sum of a vector of constant values as priors}
\usage{
calcprior(npars, N)
}
\arguments{
\item{npars}{the number of model parameters being examined by the MCMC}

\item{N}{the number of replicate parameter vectors to be returned from 
do_MCMC}
}
\value{
the sum of a vector of small constant valus to act as priors.
}
\description{
so as to include a prior probability into Bayesian
    calculations calcprior is a template for generating such priors.
    The default given here is to return a constant small number for the
    prior probability, it needs to sum to 1.0 so is the log(N) the number
    of replicates returned by do_MCMC. If non-uniform priors are 
    required write a different function and in do_MCMC point priorcalc
    at it. whatever function you define it needs to have the same input
    parameters as this calcprior, i.e. number of parameters and N. If
    something else if required then do_MCMC will need modification in
    the calculation of func1 where priorcalc is used.
}
\examples{
\dontrun{
calcprior(npars=4,N=20000)  # should give -39.61395
}
}
