% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mqmf_utils.r
\name{countgtOne}
\alias{countgtOne}
\title{countgtOne used in apply to count the number > 1 in a vector}
\usage{
countgtOne(invect)
}
\arguments{
\item{invect}{vector of values}
}
\value{
A single value of zero or the number of NAs
}
\description{
countgtOne used in apply to count the number > 1 in a vector
}
\examples{
\dontrun{
x <- matrix(trunc(runif(20)*10),nrow=4,ncol=5)
print(x)
apply(x,1,countgtone)
}
}
