% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mqmf_utils.r
\name{countgtzero}
\alias{countgtzero}
\title{countgtzero used in apply to count how many numbers are
    greater than zero in a vector}
\usage{
countgtzero(invect)
}
\arguments{
\item{invect}{vector of values}
}
\value{
A single integer counting how many numbers are > 0
}
\description{
countgtzero used in apply to count how many numbers are
    greater than zero in a vector
}
\examples{
\dontrun{
x <- matrix(trunc(runif(20)*10),nrow=4,ncol=5)
print(x)
apply(x,1,countgtzero) # count by rows
apply(x,2,countgtzero) # count by columns
}
}
