% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mqmf_utils.r
\name{countNAs}
\alias{countNAs}
\title{countNAs used in apply to count the number of NAs in a vector}
\usage{
countNAs(invect)
}
\arguments{
\item{invect}{vector of values}
}
\value{
A single value of zero or the number of NAs
}
\description{
countNAs used in apply to count the number of NAs in a vector
}
\examples{
\dontrun{
x <- matrix(trunc(runif(20)*10),nrow=4,ncol=5)
diag(x) <- NA
print(x)
apply(x,1,countNAs)
}
}
