% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mqmf_utils.r
\name{countones}
\alias{countones}
\title{countones used in apply to count the number of ones in a vector}
\usage{
countones(invect)
}
\arguments{
\item{invect}{vector of values}
}
\value{
A single value of zero or the number of ones
}
\description{
countones used in apply to count the number of ones in a vector
}
\examples{
\dontrun{
x <- matrix(trunc(runif(20)*10),nrow=4,ncol=5)
print(x)
apply(x,1,countones)
}
}
