% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mqmf_utils.r
\name{countzeros}
\alias{countzeros}
\title{countzeros used in apply to count the number of zeros in a vector}
\usage{
countzeros(invect)
}
\arguments{
\item{invect}{vector of values}
}
\value{
A single value of zero or the number of zeros
}
\description{
countzeros used in apply to count the number of zeros in a vector
}
\examples{
x <- matrix(trunc(runif(20)*10),nrow=4,ncol=5)
print(x)
apply(x,1,countzeros) # count by rows
apply(x,2,countzeros) # count by columns
}
