% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm.r
\name{displayModel}
\alias{displayModel}
\title{displayModel plots the model fit given the parameters and data}
\usage{
displayModel(inp, indat, schaefer = TRUE, extern = FALSE,
  limit = 0.2, target = 0.48, addrmse = FALSE, filename = "",
  resol = 200, fnt = 7, plotout = TRUE, vline = 0)
}
\arguments{
\item{inp}{a vector of model parameters (r,K,B0)}

\item{indat}{a matrix with at least columns 'year', 'catch', and 'cpue'}

\item{schaefer}{if TRUE, the default, then the Schaefer SPM is used. If FALSE
then an approximate Fox SPM would be used}

\item{extern}{defaults to FALSE, determines whether to plot the graphs
in a separate window}

\item{limit}{defaults to the Commonwealth limit of 0.2B0.}

\item{target}{defaults to the Commonwealth target of 0.48B0. It determines
the green line plotted on the exploitable biomass plot.}

\item{addrmse}{default is FALSE but if set TRUE this will add the loess
curve to the CPUE trend for comparison with the fitted line}

\item{filename}{default is empty. If a filename is put here a .png file
with that name will be put into the working directory.}

\item{resol}{the resolution of the png file, defaults to 200 dpi}

\item{fnt}{the font used in the plot and axes. Default=7, bold Times. Using
6 gives Times, 1 will give SansSerif, 2 = bold Sans}

\item{plotout}{should one generate a plot or only do the calculations; default
is TRUE}

\item{vline}{defaults to 0 = no action but if a year is inserted,
which should generally be between two years to designate some
change in the tme-series between two years, then a vertica line
will be added/}
}
\value{
invisibly a list of the dynamics, production curve, MSY, and Bmsy
}
\description{
displayModel takes a set of parameters and the spmdat matrix
    and plots the predicted depletion, catch, the surplus production, and
    the CPUE and the model fit to CPUE,
}
\examples{
\dontrun{
data(dataspm)
fish <- dataspm$fish
glb <- dataspm$glb
plotfishery(fish,glb)
pars <- c(0.264,4740,3064)
ans <- displayModel(pars,fish,schaefer=FALSE)
bestSP <- optim(par=pars,fn=negLLM,callfun=simpspm,indat=fish,schaefer=FALSE)
outoptim(bestSP)
ans <- displayModel(bestSP$par,fish,schaefer=FALSE)
str(ans)
}
}
