% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{do_MCMC}
\alias{do_MCMC}
\title{do_MCMC conducts an MCMC using Gibbs within Metropolis}
\usage{
do_MCMC(chains, burnin, N, thinstep, inpar, infunk, calcpred, calcdat,
  obsdat, priorcalc, scales)
}
\arguments{
\item{chains}{the number of independent MCMC chains produced}

\item{burnin}{the number of steps made before candidate parameter
vectors begin to be kept.}

\item{N}{the total number of posterior parameter vectors retained}

\item{thinstep}{the number of interations before a vector is kept; must 
be divisible by the number of parameters}

\item{inpar}{the initial parameter values (usually log-transformed)}

\item{infunk}{the function used to calculate the negative log-likelihood}

\item{calcpred}{the function used by infunk to calculate the predicted
values for comparison with obsdat.}

\item{calcdat}{the data used by calcpred to calculate the predicted
values}

\item{obsdat}{the observed data (on the same scale as the predicted)
agsinst with the predicted values are compared.}

\item{priorcalc}{a function used to calculate the prior probability for
each of the parameters in each trial parameter vector.}

\item{scales}{The re-scaling factors for each parameter to convert the
normal random deviates into +/- values on a scale that leads to
acceptance rates of between 20 and 40percent.}
}
\value{
a list of the result chains, the acceptance and rejection rates
}
\description{
do_MCMC conducts a Gibbs within Metropolis algorithm. One
   can define the number of chains, the burnin before candidate
   parameter vectors and associated probabilities are stored, the total
   number of candidate vectors to keep and the step between accepting
   a result into the posterior. One needs to input three functions: 1)
   infunk to calculate the likelihoods, 2) calcpred used within infunk
   to calculate the predicted values,used to compare with the observed
   (found in obsdat), and 3) priorcalc used to calculate the prior
   probabilities of the various parameters in each trial. N * step
   iterations are made in total although only N are stored. The jumping
   function uses random normal deviates (mean=0, sd=1) to combine with
   each parameter value (after multiplication by the specific scaling
   factor held in scales). Determination of suitable scaling values
   is generally done empirically, perhaps by trailing a small number of
   iterations to begin with.
}
\examples{
\dontrun{
 print("Still being developed")
}
}
