% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MQMF.r
\docType{data}
\name{fishdat}
\alias{fishdat}
\title{fishdat Three data objects suitable for use with MQMF.}
\format{A list of three objects
\describe{
  \item{fish}{ a data.frame containing Year, Catch, CPUE, and SE, the 
      standarderror of the CPUE estimates, if present}
  \item{glb}{ a list of global variables including maxage, M, parameters 
      for growth, weight-at-age, maturity-at-age, steepness, R0, 
      selectivity, resilience, number of ages, and the ages themselves.}
  \item{props}{ a data.frame of age, laa, waa, maa, sela, and feca}
}}
\usage{
fishdat
}
\description{
A dataset containing the fish data.frame, the glb list, and
    the props data.frame set up ready for use with simpleSA. In 
    particular it can be used with fitASPM, fitSPM (though it only 
    provides a very poor fit), run_cMSY, and DBSRA.
}
\keyword{datasets}
