% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm.r
\name{fitSPM}
\alias{fitSPM}
\title{fitSPM fits a surplus production model}
\usage{
fitSPM(pars, fish, schaefer = TRUE, maxiter = 1000)
}
\arguments{
\item{pars}{the initial parameter values to start the search for the optimum}

\item{fish}{the matrix containing the fishery data 'year', 'catch', and
'cpue' as a minimum.}

\item{schaefer}{if TRUE, the default, then simpspm is used to fit the
Schaefer model. If FALSE then the Fox model is fitted setting the p
parameter to 1e-08 inside simpspm.}

\item{maxiter}{the maximum number of iterations to be used in each optim run}
}
\value{
an optim output object as a list
}
\description{
fitSPM fits a surplus production model (either Schaefer or Fox)
    by applying optim twice. Being automated it is recommended that this only
    be used once plausible initial parameters have been identified (through
    rules of thumb or trial and error). It uses negLLM to apply a negative
    log-likelihood, assuming log-normal ressidual errors. The output object
    is the usual object output from optim. The $par values can be used in
    displayModel to plot the outcome, or in bootspm to conduct bootstrap
    sampling of the residuals from the CPUE model fit to gain an appreciation
    of any uncerainty in the analysis. It uses the magnitude function to set
    the values of the parscale parameters.
}
\examples{
\dontrun{
 data(dataspm)
 pars <- c(r=0.2,K=6000,Binit=2800)
 ans <- fitSPM(pars,dataspm$fish,schaefer=TRUE,maxiter=1000)
 outoptim(ans)
 ansF <- fitSPM(pars,dataspm$fish,schaefer=FALSE,maxiter=1000)
 outoptim(ansF)
}
}
