% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mqmf_utils.r
\name{freqMean}
\alias{freqMean}
\title{freqMean calculates the mean and stdev of count data}
\usage{
freqMean(values, infreqs)
}
\arguments{
\item{values}{the values for which there are counts}

\item{infreqs}{the counts for each of the values empty cells can be
either 0 or NA}
}
\value{
a vector containing the mean and st.dev.
}
\description{
freqMean calculates the mean and stdev of count data
    it requires both the values and their associated counts and
    return a vector of two numbers.
}
\examples{
vals <- c(1,2,3,4,5)
counts <- c(3,NA,7,4,2)
freqMean(vals,counts)  # should give 3.125 and 1.258306
}
