% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mqmf_utils.r
\name{getname}
\alias{getname}
\title{getname returns the name of a variable as character}
\usage{
getname(x)
}
\arguments{
\item{x}{any variable whose name is wanted as a character string}
}
\value{
a character string with the name of input variable
}
\description{
getname runs 'deparse(substitute(x))' to get the
    name of the input variable. Saves remembering the syntax
}
\examples{
a_variable <- c(1,2,3,4,5,6,7,8)
getname(a_variable)
}
