% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mqmf_utils.r
\name{getseed}
\alias{getseed}
\title{getseed generates a random number seed}
\usage{
getseed()
}
\value{
a 6 digits integer
}
\description{
getseed generates a random number seed for use within
    set.seed. It produces a 6 digit integer from the Sys.time. This
    is effectively what the system does. As the help on set.seed says
    Initially, there is no seed; a new one is created from the current 
    time and the process ID when one is required. Here in getseed we
    do not use the process ID so the process is not identical but this
    at least allows the set.seed value to be stored should the need to 
    repeat a set of simulations arise.
}
\examples{
\dontrun{
seed <- getseed()
set.seed(seed)
rnorm(5)
set.seed(seed)
rnorm(5)
}
}
