% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mqmf_utils.r
\name{getsingle}
\alias{getsingle}
\title{getsingle extracts one number from an input line of characters}
\usage{
getsingle(inline, sep = ",")
}
\arguments{
\item{inline}{the line of text, usually taken after using readLines}

\item{sep}{the separator used to divide the numbers from descriptive text.
defaults to a comma.}
}
\value{
a single number
}
\description{
getsingle splits up a text line and translates the first non-
    empty character string into a number.
}
\examples{
\dontrun{
x <- "12.3 , this is a number"
y <- "21.3 # 22.3 # here are two numbers"
getsingle(x)
getsingle(y,sep="#")
getsingle(y) # be sure to get the separator correct
}
}
