% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mqmf_utils.r
\name{\%ni\%}
\alias{\%ni\%}
\title{'\%ni\%' identifies which element in x is NOT in y}
\usage{
x \%ni\% y
}
\arguments{
\item{x}{a vector of elements which can be numeric or character}

\item{y}{a vector of elements which can be numeric or character}
}
\description{
'\%ni\%' identifies which element in x is NOT in y
}
