% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MQMF.r
\docType{data}
\name{invert}
\alias{invert}
\title{invert data derived from a trawl caught invertebrate fishery.}
\format{A data.frame of fishery dependent data
\describe{
  \item{year}{literally the year in which the fishery occurred}
  \item{catch}{the catch landed intonnes; there is little or no
     discarding}
  \item{cpue}{the standardized cpue from the fishery, this omits
     major changes in the fishery such as the introduction of
     quotas in 1992 and a major structural adjustment at the end of
     2006}
  \item{se}{the standard error of the standardized cpue. This is
     very small as there are usually a large number of observations
     but that ignores the between year variation.}
  \item{geom}{the nominal geometric mean cpue}
  \item{vessels}{the number of vessels reporting catches in each year}
  \item{records}{the number of records each year.}
}}
\usage{
invert
}
\description{
A dataset containing the invert data.frame as a 31 x 7
    matrix. The invert data.frame has both standardized cpue as
    well as the unstandardized geometric mean cpue, labelled 'geom'.
    There is also columns of the catch, the standard error of the
    standardized cpue (too small to be useful in stock assessments),
    the number of vessels reporting each year, and the number of
    records each year. Very difficult to obtain a plausible solution
    when trying to apply a spm model.
}
\examples{
 \dontrun{
 data(invert)
 invert
}
}
\keyword{datasets}
