% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MQMF.r
\docType{data}
\name{kimura}
\alias{kimura}
\title{kimura simplified length-at-age data.}
\format{A data.frame of simplified length-at-age data
\describe{
  \item{age}{the estimated age of fish}
  \item{length}{the mean length of fish at each age}
}}
\usage{
kimura
}
\description{
A dataset containing the age and length data from
    Kimura, D.K. (1980) Likelihood methods for the von Bertalanffy
    growth curve. Fishery Bulletin, 77: 765-776. There are only 11
    observations because the lengths are averages, which, these
    days would be unusual as computers and software can easily
    handle many hundreds of observations. But things were more
    difficult in 1980!
}
\examples{
 \dontrun{
 data(kimura)
 kimura
}
}
\keyword{datasets}
