% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisheries.r
\name{MaA}
\alias{MaA}
\title{MaA alternative logistic function commonly used for maturity}
\usage{
MaA(ina, inb, depend)
}
\arguments{
\item{ina}{is the intercept of the exponential function}

\item{inb}{is the gradient of the exponential function}

\item{depend}{is a vector of lengths/ages for which the logistic 
maturity value will be calculated}
}
\value{
A vector of length(depend) with predicted maturity values
}
\description{
MaA a function exp(a+bxdepend)/(1+exp(a+bxdepend)),
    which can also be expressed as 1/(1+(1/exp(a + b x depend))) to
    for a symmetric logistic curve. This has the property that the 
    SM50 = -a/b and the interquartile distance is 2.Ln(3)/b.
}
\examples{
a <- -14.383
b <- 0.146017
lens <- seq(2,210,2)
round(MaA(a,b,depend=lens),5) # length based
round(MaA(-2.5,0.95,0:25),5)   # age based
}
