% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mqmf_utils.r
\name{magnitude}
\alias{magnitude}
\title{magnitude returns the magnitude of numbers in base 10}
\usage{
magnitude(x)
}
\arguments{
\item{x}{the vector of numbers (parameters) whose magnitudes are
needed}
}
\value{
a vector of magnitudes
}
\description{
magnitude is useful when using an
    optimizer such as optim, which uses a parscale parameter.
    magnitude can determine the respective parscale value for each
    parameter value.
}
\examples{
  x <- c(0,0.03,0.3,3,30,300,3000)
  magnitude(x)
}
