% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisheries.r
\name{negnormL}
\alias{negnormL}
\title{negnormL an alternative -log-likelihood for normal errors}
\usage{
negnormL(par, funk, indat, obs = "deltal", ...)
}
\arguments{
\item{par}{the vector of parameters, with sigma, the standard
deviation of the normal random deviates at the end.}

\item{funk}{the funk needed to generate the predicted values}

\item{indat}{the data.frame or matrix containing the obs and the 
independent variable used by funk}

\item{obs}{identifies the column name or column number that contains 
the observed data for comparison with the predicted from funk}

\item{...}{the standard R for including extra parameters needed by
funk but without having to be explicitly defined.}
}
\value{
the negative log-likelihood, for use in an optimizer, eg nlm
}
\description{
negnormL is an alternative to negNLL to produce
    -ve log-likelihoods for nromal random errors. The two functions
    differ by how they reference the independent and dependent 
    variables. In negnormL only the observed (for comparison with the
    predicted) is identified. This is to allow for different column
    headings in one's data.frames
}
\examples{
\dontrun{
 data(blackisland)
 param <- c(Linf=173.65,K=0.2666,sigma=3.6)
 negnormL(par=param,funk=fabens,indat=blackisland,obs="deltal",
          initL="len1",delT="deltat")   # should be 291.6809
 param2 <- c(21.03,130.94,40.65,3.162)  
 negnormL(par=param2,funk=invl,indat=blackisland,obs="deltal",
          initL="len1",delT="deltat")  # should be 277.5663 
}
}
