% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MQMF.r
\docType{data}
\name{npf}
\alias{npf}
\title{npf is the brown and grooved tiger prawn fishery data from NPF}
\format{A matrix of fisheries data
\describe{
  \item{year}{the fishing year from 1970 - 1998}
  \item{catch}{the total annual catch, tonnes }
  \item{cpue}{the catch tonnes per standard day, a ratio cpue}
}}
\usage{
npf
}
\description{
npf is fishery data from teh Northern Prawn Fishery of 
    Australia. It represents the combined catches of the brown tiger
    prawn (Penaeus esculentus) and the grooved tiger prawn (Penaeus
    semisulcatus) from 1970 to 1998. The data are derived from Pownall,
    P.C. (ed.) (1994) Australia's Northern Prawn Fishery: The first 25 
    years. NPF25. Cleveland, Austrlaia.
}
\examples{
 \dontrun{
 data(npf)
 npf
}
}
\keyword{datasets}
