% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mqmf_utils.r
\name{penalty0}
\alias{penalty0}
\title{penalty0 enables the adding of a large penalty as one approaches 0.0}
\usage{
penalty0(x)
}
\arguments{
\item{x}{the parameter value that potentially incurs a penalty}
}
\value{
a single value as a penalty to be added to a Log-Likelihood or SSQ
}
\description{
penalty0 allows for the option of adding a large penalty as
    a parameter approaches 0.0 . See negLL1 for example code that 
    contains such a parameter. For example, when
    fitting an spm sometimes the optimal mathematical model fit can occur
    by depressing the r value to 0 or even go negative. Input values 
    < 0.006 begin to generate large values as one goes smaller. The
    examples below illustrate this.
}
\examples{
  penalty0(0.5)
  penalty0(0.1)
  penalty0(0.01)
  penalty0(0.005)
}
