% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfuns.R
\name{plot.dlpop}
\alias{plot.dlpop}
\title{plot.dlpop an S3 method for plotting dlpop objects}
\usage{
\method{plot}{dlpop}(x, main = "", cex = 0.9, font = 7, ...)
}
\arguments{
\item{x}{a dlpop matrix containing year, nt, and nt1 as columns}

\item{main}{defines text to print across top of plot with the
intention of including the parameter values, default=""}

\item{cex}{the size of the fonts used, default=0.9}

\item{font}{the font used default=7 (bold serif) 1 = non-serif,
2 = bold non-serif, 6 = serif}

\item{...}{ready for extra graphical parameters}
}
\value{
it returns the dlpop matrix invisibly
}
\description{
plot.dlpop an S3 method for plotting dlpop objects
}
\examples{
\dontrun{
  r <- 2.5; K <- 1000.0; N0 <- 50; Ct <- 0.0; yrs <- 50; p <- 1
  pop <- discretelogistic(r=r,K=K,N0=N0,Ct=Ct,Yrs=yrs,p=p)
  plot(pop,main=paste0("r=",r,"  K=",K,"  Ct=",Ct, "  N0=",N0,
                       "  p=",p),cex=0.85,font=7)
}
}
