% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm.r
\name{plot.spmdat}
\alias{plot.spmdat}
\title{plot.spmdat S3 method that plots an spmdat data set}
\usage{
\method{plot}{spmdat}(x, ...)
}
\arguments{
\item{x}{a data set that should contain 'year', 'catch', and
'cpue'}

\item{...}{other arguments}
}
\value{
nothing, but it does generate a new plot
}
\description{
plot.spmdat is an S3 method that plots an spmdat
    data set. It plots the catch and CPUE against time
}
\examples{
\dontrun{
yrs <- 2000:2010
catches <- rnorm(length(yrs),mean=150,sd=5)
ce <- rnorm(length(yrs),mean=20,sd=1)
makemat <- cbind(year=yrs,catch=catches,cpue=ce)
dat <- makespmdata(makemat)
plot(dat)
}
}
