% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mqmf_utils.r
\name{plot1}
\alias{plot1}
\title{plot1 a simple way to plot an xy line plot}
\usage{
plot1(x, y, xlabel = "", ylabel = "", type = "l", usefont = 7,
  cex = 0.85, maxy = 0, defpar = TRUE)
}
\arguments{
\item{x}{The single vector of x data}

\item{y}{the single vector of y data. If more are required they can
be added spearately after calling plot1.}

\item{xlabel}{the label fot the x-axis, defaults to empty}

\item{ylabel}{the label fot the y-axis, defaults to empty}

\item{type}{the type of plot "l" is for line, the default, "p" is
points. If you want both plot a line and add points afterwards.}

\item{usefont}{which font to use, defaults to 7 which is Times bold}

\item{cex}{the size of the fonts used. defaults to 0.85}

\item{maxy}{defaults to 0, which does nothing. If a value is given
then this value is used rather than estimating from the input y}

\item{defpar}{if TRUE then plot1 will declare a par statement. If false it
will expect one outside the function. In this way plot1 can be
used when plotting multiple graphs, perhaps as mfrow=c(2,2)}
}
\value{
nothing but it does plot a graph and changes the par setting
}
\description{
plot1 provides a quick way to plot out a single xy
    line plot. It can be used with plotprep to generate a plot
    outside of Rstudio or by itself to generate one within Rstudio.
    It uses a standard par setup and permits custom labels, font,
    and font size (cex). It checks the spread of y and if a ymax is
    not given in the parameters finds the ymax and checks to see if
    y goes negative in which case it uses getmin, so the
    y-axis is set to 0 - ymax or ymin - ymax
}
\examples{
\dontrun{
x <- rnorm(20,mean=5,sd=1)
plot1(x,x,xlabel="x-values",ylabel="yvalues")
}
}
