% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mqmf_utils.r
\name{plot2}
\alias{plot2}
\title{plot2 prepares for 2 plots one on top of the other}
\usage{
plot2(width = 6, height = 5, usefont = 7, cex = 0.9,
  newdev = TRUE, titleX = FALSE, titleY = FALSE, rows = TRUE)
}
\arguments{
\item{width}{defaults to 6 inches = 15.24cm - width of plot}

\item{height}{defaults to 5 inches = 12.7 cm - height of plot}

\item{usefont}{default is 7 (bold Times); 1 = sans serif, 2 = sans serif bold}

\item{cex}{default is 1.0, but is used to alter the standard size of text}

\item{newdev}{make a new graphics device or reuse a previously defined one;
defaults to TRUE}

\item{titleX}{if TRUE then standard x-axis titles are expected for
plot and space is left appropriately, if FALSE then mtext is
expected to be used instead. Defaults to FALSE}

\item{titleY}{if TRUE then standard y-axis titles are expected for
plot and space is left appropriately, if FALSE then mtext is
expected to be used instead.  Defaults to FALSE}

\item{rows}{if rows = TRUE then 2 plots one above the other, if
FALSE then 2 plots side by side}
}
\value{
nothing but it prepares a device for 2 plots
}
\description{
plot2 prepares for 2 plots one on top of the other.
    it includes the option of having separate titles for each
    plot or using mtext to cover off on both
}
\examples{
\dontrun{
x1 <- rnorm(1000,mean=0,sd=1.0)
x2 <- rnorm(1000,mean=0,sd=1.0)
plot2(rows=FALSE)
hist(x1,breaks=30,main="",xlab="",ylab="",col=2)
hist(x2,breaks=30,main="",xlab="",ylab="",col=2)
mtext("Frequency",side=2,outer=TRUE,line=0.0,font=7,cex=1.0)
mtext("Normal Random Variates",side=1,outer=TRUE,line=0.0,font=7,cex=1.0)
}
}
