% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm.r
\name{plotModel}
\alias{plotModel}
\title{plotModel plots the model fit given the parameters and data}
\usage{
plotModel(inp, indat, schaefer = TRUE, extern = FALSE, limit = 0.2,
  target = 0.48, addrmse = FALSE, filename = "", resol = 200,
  fnt = 7, plotout = TRUE, vline = 0, plotprod = TRUE)
}
\arguments{
\item{inp}{a vector of model parameters at least (r,K,B0)}

\item{indat}{a matrix with at least columns 'year', 'catch', and 'cpue'}

\item{schaefer}{if TRUE, the default, then the Schaefer SPM is used. 
If FALSE then an approximate Fox SPM would be used}

\item{extern}{defaults to FALSE, determines whether to plot the graphs
in a separate window rather than within RStudio or an Rmd file}

\item{limit}{defaults to the Commonwealth limit of 0.2B0.}

\item{target}{defaults to the Commonwealth target of 0.48B0. Determines
the green line plotted on the exploitable biomass plot.}

\item{addrmse}{default is FALSE but if set TRUE this will add the loess
curve to the CPUE trend for comparison with the fitted line}

\item{filename}{default is empty "". If a filename is put here a 
.png file with that name will be put into the working directory.}

\item{resol}{the resolution of the png file, defaults to 200 dpi}

\item{fnt}{the font used in the plot and axes. Default=7, bold Times. 
Using 6 gives Times, 1 will give SansSerif, 2 = bold Sans}

\item{plotout}{should one generate a plot or only do the calculations; 
default is TRUE}

\item{vline}{defaults to 0 = no action but if a year is inserted,
which should generally be between two years to designate some
change in the time-series between two years, then a vertical line
will be added on year-based plots}

\item{plotprod}{if TRUE, the default, the productivity curves are also
plotted.}
}
\value{
invisibly a list of the dynamics, production curve, MSY, and Bmsy
}
\description{
plotModel takes a parameter set and the spmdat matrix
    and plots the predicted depletion, catch, the CPUE and the model 
    fit to CPUE. If plotprod=TRUE, it also plots the productivity curves
}
\examples{
\dontrun{
data(dataspm)
fish <- dataspm$fish
glb <- dataspm$glb
data(dataspm) 
pars <- log(c(0.264,4740,3064,0.2))
ans <- plotModel(pars,fish,schaefer=TRUE,extern=TRUE,plotprod=TRUE)
best <- nlm(negLL,pars,funk=simpspm,indat=fish,logobs=log(fish[,"cpue"]),
           typsize=magnitude(pars),steptol=1e-05)
outfit(best,backtransform = TRUE)
ans <- plotModel(best$estimate,fish,schaefer=TRUE,extern=TRUE,
                 plotprod=FALSE,addrmse=TRUE)
str(ans)
}
}
