% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mqmf_utils.r
\name{quants}
\alias{quants}
\title{quants used in apply to estimate quantiles across a vector}
\usage{
quants(invect)
}
\arguments{
\item{invect}{vector of values}
}
\value{
a vector of the c(0.025,0.05,0.5,0.95,0.975) quantiles
}
\description{
quants used in 'apply' to estimate quantiles across a vector
}
\examples{
cat("Example still be made /n")
}
