% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mqmf_utils.r
\name{readdata}
\alias{readdata}
\title{readdata reads in a standard format data file}
\usage{
readdata(filename, property = FALSE, verbose = TRUE)
}
\arguments{
\item{filename}{the filename (including the full path if required) containing
the data in the standard format.}

\item{property}{does the data include a table of length-at-age, maturity-at-
age, weight-at-age, fecundity-at-age, and that kind of thing (see the
standard format for a list of data.)}

\item{verbose}{Default = TRUE, which prints out details as data is read in.}
}
\value{
a list of three objects, fish includes the Year, Catch, CPUE, and SE
    of the CPUE, glb containing an array of biological properties that can
    be global, and props containing age, laa, waa, maa, sela.
}
\description{
readdata reads in a standard format data file. An example of the
    standard format is generated by using the function dataTemplate, which
    generates an example datafile which can be used to demonstrate the
    methods present in FMR, or can be used as a template to edit and
    input a new or different dataset.
}
\examples{
\dontrun{
dataTemplate(filename="test.csv", title="A test of the functions")
ans <- readdata("test.csv",property=FALSE)
str(ans)
}             #
}
