% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mqmf_utils.r
\name{removeEmpty}
\alias{removeEmpty}
\title{removeEmpty removes empty strings from a vector of strings}
\usage{
removeEmpty(invect)
}
\arguments{
\item{invect}{a vector of input strings, possibly containing empty strings}
}
\value{
a possibly NULL vector of strings
}
\description{
removeEmpty removes empty strings from a vector of 
    strings. Such spaces often created by spurious commas at the 
    end of lines. It also removes strings made up only of spaces 
    and removes spaces from inside of inidivdual chunks of text. 
    So, should be useful when reading in data from a custom csv 
    file when parsing different formats
}
\examples{
x <- c("1","","2","","   ","3"," ","4","","a string","end")
x
length(x)
length(removeEmpty(x))
removeEmpty(x)
}
