% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisheries.r
\name{simpspm}
\alias{simpspm}
\title{simpspm simply calculates the predicted log(CE) for an SPM}
\usage{
simpspm(pars, indat, schaefer = TRUE, depleted = TRUE, year = "year",
  cats = "catch", index = "cpue")
}
\arguments{
\item{pars}{the parameters of the SPM are either c(r, K, Binit, sigma),
Binit is required if the fishery data starts after the stock has been
depleted. Each parameter must be log-transformed for improved model
stability and is transformed inside simpspm.}

\item{indat}{the data which needs to include year, catch, and cpue.}

\item{schaefer}{a logical value determining whether the spm is to be a
simple Schaefer model (p=1) or approximately a Fox model (p=1e-08). The
default is TRUE = Schaefer model}

\item{depleted}{default = TRUE; implies the fishery was already depleted
once fishery data started to be collected. Implies that there needs
to be a Binit in third position among the input parameters}

\item{year}{the column name within indat containing the years}

\item{cats}{the column name within indat containing the catches}

\item{index}{the column name within indat containing the cpue.}
}
\value{
a vector of length nyrs of log(cpue)
}
\description{
simpspm calculates the predicted log(CPUE) for an SPM model. 
    It set the Polacheck et al parameter 'p' depending on the schaefer
    term, and this determines the asymmetry of the production curve.
    If p = 1.0 then the SPM is the Schaefer model, if it is 1e-8 it
    approximates the Fox model. The output of log(CPUE) is to simplfy
    the use of log-normal residual errors or likelihoods. This function 
    is designed for data consisting of only a single cpue time-series.
}
\examples{
\dontrun{
 data(abdat)
 fish <- abdat$fish
 colnames(fish) <- tolower(colnames(fish))
 param <- log(c(r=0.4,K=9400,Binit=3400,sigma=0.05))
 predCE <- simpspm(pars=param,fish)
 cbind(fish,exp(predCE))
} 
}
