% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm.r
\name{simpspmO}
\alias{simpspmO}
\title{simpspmO simply calculates the predicted CE for an SPM}
\usage{
simpspmO(
  par,
  indat,
  schaefer = TRUE,
  year = "year",
  cats = "catch",
  index = "cpue"
)
}
\arguments{
\item{par}{the parameters of the SPM = r, K, a q for each column of cpue,
a sigma for each cpue, and Binit if fishery depleted to start with. 
Each parameter is in log space and is transformed inside simpspm}

\item{indat}{the data which needs to include year, catch, and cpue. The
latter should have a separate column for each fleet, with a column 
name beginning with cpue or whatever name you put in index (see below)
for example cpue1, cpue2,etc.}

\item{schaefer}{a logical value determining whether the spm is to be a
simple Schaefer model (p=1) or approximately a Fox model (p=1e-08). 
The default is TRUE}

\item{year}{the column name within indat containing the years}

\item{cats}{the cloumn name within indat containing the catches}

\item{index}{the prefix in the column names given to the indices of 
relative abundance used, perhaps 'cpue' as in cpueTW, cpueAL, etc. 
grep is used to search for columns containing this prefix to 
identify whether there are more than one column of cpue data.}
}
\value{
a vector or matrix of nyrs of the predicted CPUE
}
\description{
simpspmO calculates the predicted CPUE for an SPM 
    model. It generates the predicted cpue without log-transformatio
}
\examples{
\dontrun{
data(dataspm)
dataspm
colnames(dataspm) <- tolower(colnames(dataspm))
pars <- log(c(r=0.242,K=5170,Binit=2840,q=0.0002))
predCE <- simpspmO(pars,dataspm)
cbind(dataspm[,"year"],dataspm[,"cpue"],predCE)
}
}
