% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mqmf_utils.r
\name{ssq}
\alias{ssq}
\title{ssq a generalized function for summing squared residuals}
\usage{
ssq(par, funk, independent, observed)
}
\arguments{
\item{par}{a vector of parameters used by funk}

\item{funk}{a function that uses par and indep to calculate
predicted values to compare with the dep values}

\item{independent}{a vector containing the independent variable
(x-axis) upon which the observed variable is assumed to be
dependent.}

\item{observed}{a vector containing the observed data (y-axis)}
}
\value{
a single number (scaler) that is the sum of squared
    residuals between the dep values and those calculated by funk
}
\description{
ssq is a generalized function for summing squared
    residuals which is designed for ease of use in nlm (or
    optim, or nlminb). NAs are removed. It assumes the input of
    a predefined function 'funk' that will calculate predicted
    values of a 'dependent' variable from a vector of 'independent'
    values, for which one has observations. The predicted values
    are compared with the observed 'dependent data, and the
    resulting SSQ returned.
}
\examples{
\dontrun{
   print("an example being developed")
}
}
