% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm.r
\name{ssqL}
\alias{ssqL}
\title{ssqL a function for summing log-transformed squared residuals}
\usage{
ssqL(inp, indat, callfun)
}
\arguments{
\item{inp}{a vector of parameters used by the input function 'funk'}

\item{indat}{the data set containing the 'year', 'catch', and 'cpue'}

\item{callfun}{the function that calculates the predicted values from
the input data}
}
\value{
a single number (scaler) that is the sum of squared
    residuals between the obs cpue and those calculated by infun
}
\description{
ssqL is a generalized function for summing squared
    residuals which is designed for ease of use in optim or nlm, etc.
    This contrasts with ssq in that it estimates log-transformed
    data and predicted values
}
\examples{
\dontrun{
data(dataspm)
fish <- dataspm$fish
fish
colnames(fish) <- tolower(colnames(fish))
pars <- c(r=0.2,K=6000,Binit=2800)
ssq(pars,fish,simpfox)  # should be 1.713938
ssq(pars,fish,simpspm)  # should be 0.8856442
}
}
