% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MQMF.r
\docType{data}
\name{tigers}
\alias{tigers}
\title{tigers is spawning biomass and subsequent recruitment data}
\format{A data.frame of spawning recruitment data
\describe{
  \item{Spawn}{the estimated spawning biomass in a year}
  \item{Recruit}{estimated recruitment from the biomass in each year}
}}
\usage{
tigers
}
\description{
tigers is a dataset from Penn, J.W. and N. Caputi (1986)
    Spawning stock-recuitment relationships and environmental
    influences on the tiger prawn (_Penaeus esculentus_) fishery in
    Exmouth Gulf, Western Australia. _Australian Journal of Marine
    and Freshwater Research_ __37__: 491-505. It is only 14 rows of
    data with a column of Spawn and Recruit, as a data.frame
}
\examples{
 \dontrun{
 data(tigers)
 tigers
}
}
\keyword{datasets}
