% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisheries.r
\name{bce}
\alias{bce}
\title{bce Baranov catch equation}
\usage{
bce(M, Fat, Nt, ages)
}
\arguments{
\item{M}{instantaneous rate of natural mortality, assumed constant}

\item{Fat}{a vector of age-specific instantaneous rates of fishing
mortality over the time period t, this is usually estimated by
multiplying the fully selected F by the selectivity/availability
at age.}

\item{Nt}{The population numbers at the start of time t}

\item{ages}{the ages 0:maxage used in the calculations}
}
\value{
a matrix of numbers-, natural mortality-, and catch-at-age
}
\description{
bce the Baranov Catch Equation. The total kill is made
    up of fish being killed by fishing and other dying naturally.
    We use the bce to estimate the catch (those killed by fishing).
    The problem is that some fish that would be expected to die 
    naturally can be expected to be caught and killed by fishing
    so estimating the catch is slightly more complex than Nt x Ht.
    It is invariably better to use the Baranov Catch Equation.
}
\examples{
\dontrun{
Ft <- -log(1 - 0.2) # harvest rate of 0.2
M <- 0.12
Nt <- 1000
bce(M,Ft,Nt)   # should give 188.8862
}
}
