% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mqmf_utils.r
\name{countgtone}
\alias{countgtone}
\title{countgtone used in apply to count the number > 1 in a vector}
\usage{
countgtone(invect)
}
\arguments{
\item{invect}{vector of values}
}
\value{
the number of values in the vector > 1
}
\description{
countgtone used in the base function 'apply', or 'tapply' to 
    count the number > 1 in a vector
}
\examples{
x <- matrix(trunc(runif(20)*10),nrow=4,ncol=5)
print(x)
apply(x,1,countgtone)
}
