% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisheries.r
\name{discretelogistic}
\alias{discretelogistic}
\title{discretelogistic example box 2.2 Discrete logistic model}
\usage{
discretelogistic(r = 0.5, K = 1000, N0 = 50, Ct = 0, Yrs = 50,
  p = 1, outplot = TRUE, title = "")
}
\arguments{
\item{r}{intrinsic rate of population increase; default = 0.5}

\item{K}{carrying capacity; default = 1000}

\item{N0}{Initial population size; default = 50}

\item{Ct}{annual catch default = 0.0}

\item{Yrs}{years of population growth}

\item{p}{the asymmetry parameter. the default value of 1.0 leads
to the classical Schaefer model. A value of 1e-08 approximates
the Fox model where the mode is skewed to the left of centre.}

\item{outplot}{default = TRUE; should the dynamics be plotted?}

\item{title}{If using the default = "" it is filled with r, K, Ct,
N0, and p. But this can be replaced with text if required.}
}
\value{
if outplot = TRUE it generates a plot and alters default
    par settings. In addition, invisibly returns a matrix of year,
    Nt, and Nt+1
}
\description{
discretelogistic ex. box 2.2 Discrete logistic model
}
\examples{
\dontrun{
  discretelogistic(0.5,1000.0,25,0.0,50)
  ans <- discretelogistic(r=2.55,K=1000.0,N0=100,Ct=95.0,Yrs=100)
  head(ans,20)
}
}
