% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisheries.r
\name{Gz}
\alias{Gz}
\title{Gz calculates the predicted Gompertz length at age growth curve}
\usage{
Gz(p, ages)
}
\arguments{
\item{ages}{is a vector of ages; could be a single number}

\item{par}{is a vector the first three cells of which are a, b, c, for
the Gz curve.}
}
\value{
a vector of predicted lengths for the vector of ages in 'ages'
}
\description{
Gz calculates length at age for the Gompertz curve.
}
\examples{
\dontrun{
ages <- seq(0,20,1)   # sigma is ignored here
pars <- c(a=26.0,b=0.7,c=-0.5,sigma=1.0) # a, b, c, sigma; 
cbind(ages,Gz(pars,ages))
}
}
