% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mqmf_utils.r
\name{iscol}
\alias{iscol}
\title{incol is a utility to determine is a column is present in a matrix}
\usage{
iscol(incol = "year", inmat)
}
\arguments{
\item{incol}{the name of the column; defaults to "year" as an example}

\item{inmat}{the matrix or data.frame within which to search for incol}
}
\value{
TRUE or FALSE
}
\description{
incol is a utility to determine whether a names columns is
    present in a given matrix or data.frame.
}
\examples{
\dontrun{
test <- matrix(c(1,2,3,4),nrow=2,ncol=2,dimnames=list(1:2,c("year","Catch")))
print(test)
iscol("year",test)
iscol("Catch",test)
iscol("catch",test)
iscol("ages",test)
}
}
