% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{LatA}
\alias{LatA}
\title{LatA Length at age for 358 female fish}
\format{A data.frame with 358 rows and 3 variables:
\describe{
  \item{Sex}{gender coded as 1 = males and 2 = females, this is an 
      unusual coding as very often one finds females=1 and males=2. 
      This shows that simple codes cannot be taken for granted and
      can therefore be dangerous so watch out for them}
  \item{Length}{Fork length of the fish, in cms}
  \item{Age}{Estimated age from otolith reading}
}}
\source{
This data is a small sub-set of very many more samples collected 
    across species and years by the many excellent people running the 
    Integrated Stock Monitoring Program in the Australian South East 
    Fishery over the years of its existence. Thanks to the Australian
    Fisheries Management Authority for permission to use this data.
}
\description{
A data.frame containing an index number, the estimated 
    age for 358 redfish (\emph{Centroberyx affinis}) from eastern Australia 
    sampled in 1997, all from a single port.
}
\section{Subjects}{

 \itemize{
  \item Estimating individual growth from length-at-age data
  \item von Bertalanffy growth curve
  \item Gompertz growth curve
  \item Michaelis-Menton curve used as a growth curve 
 }
}

\examples{
\dontrun{
  data(LatA)
  pars <- c(27.0,0.15,-2.0) # von Bertalanffy
  bestvB <- nlm(f=ssq,funk=vB,observed=LatA$length,p=pars,
                ages=LatA$age,typsize=magnitude(pars),iterlim=1000)
                outfit(bestvB,backtran=FALSE,title="vB")
}
}
