% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MQMF.r
\docType{data}
\name{LatA}
\alias{LatA}
\title{LatA Length at age for 576 fish including both males and females.}
\format{A data frame with 576 rows and 3 variables:
\describe{
  \item{Sex}{gender coded as 1 = males and 2 = females}
  \item{Length}{Fork length of the fish, in cms}
  \item{Age}{Estimated age from otolith reading}
}}
\usage{
LatA
}
\description{
A dataset containing the sex, the length and the estimated age
   for 576 redfish (Centroberyx affinis) from eastern Australia sampled in 1997,
   all from a single port.
}
\keyword{datasets}
