% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MQMF.r
\docType{data}
\name{minnow}
\alias{minnow}
\title{minnow is individual growth data for use with growth curves}
\format{A data.frame of length-at-age data
\describe{
  \item{week}{the week of sampling for lengths}
  \item{length}{the mean length in the corresponding week in mm}
}}
\usage{
minnow
}
\description{
minnow is a dataset derived from Pitcher & Macdonald 
    (1973) for use when fitting growth curves, especially seasonal 
    growth curves. The data exhibit increases and decreases because 
    these are mean lengths rather than individual measurements 
    (which would be more typically used these days.). The data have 
    been read off the graphs within the paper as it is not reported 
    explicitly and are therefore only approximate, but will do for
    our purposes (but expect different parameters to those shown 
    in the original paper.
}
\examples{
 \dontrun{
 data(minnow)
 minnow
}
}
\references{
Pitcher, T.J., and P.D M. MacDonald. (1973) Two models 
    of seasonal growth. Journal of Applied Ecology 10:599–606.
}
\keyword{datasets}
