% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisheries.r
\name{negLL}
\alias{negLL}
\title{negLL calculate log-normal log-likelihoods}
\usage{
negLL(pars, funk, logobs, ...)
}
\arguments{
\item{pars}{the log-transformed parameters to be used in the funk for
calculating the log of the predicted values against which the log
observed values will be compared}

\item{funk}{the function used to calculate the log-predicted values 
of whatever variable is being used (eg. cpue, catches, etc.)}

\item{logobs}{the observed values log-transformed ready for 
comparison with the log-predicted values from funk and pars.}

\item{...}{required to allow funk to access its other arguments 
without having to explicitly declare them in negLL. In the 
example below, indat is passed via the ...}
}
\value{
the negative log-likelihood using log-normal errors.
}
\description{
negLL calculates log-normal negative log-likelihoods. It
    expects the input parameters to be log-transformed, so the funk 
    used to calculate the log or the predicted values also needs to 
    expect log-transformed parameters. In addition, it checks that 
    there are no missing data (NA) within the input observed 
    log-transformed data. If there are it uses only those records 
    for which there are values.
}
\examples{
\dontrun{
data(abdat)
param <- log(c(r= 0.42,K=9400,Binit=3400,sigma=0.05))
negLL(pars=param,funk=simpspm,logobs=log(abdat[,"cpue"]),indat=abdat)
}
}
