% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mqmf_utils.r
\name{negNLL}
\alias{negNLL}
\title{negNLL  -ve log-likelihood for normally distributed variables}
\usage{
negNLL(pars, funk, independent, observed)
}
\arguments{
\item{pars}{a vector containing the parameters being used in funk, plus
an extra sigma which is the standard deviation of the normal random
likelihoods in dnorm}

\item{funk}{the function name that calculates the predicted values from
the independent values}

\item{independent}{the x-axis variable, that which in the model gives rise
to the predicted values to compare with the observed 'dependent' variable}

\item{observed}{the observed values of the values that the model will
predict for each of the independent values.}
}
\value{
the sum of the negative log-likelihoods using a normal PDF
}
\description{
negNLL - Calculates the negative log-likelihood for
    normally distributed variables. It assumes the presence of a function
    'funk' that will calculate predicted values of a 'dependent' variable
    from a vector of 'independent' values.
    Using negNL takes approximately 0.4 milliseconds less time than the
    same analysis conducted using negNLP; this was determined using the
    microbenchmark package.
}
\examples{
\dontrun{
 data(kimura)
 pars <- c(Linf=56.0,K=0.4,t0=0.2,sigma=1.5)
 negNLL(pars,vB,independent=kimura[,"age"],observed=kimura[,"length"])
 # should be 19.20821
}
}
