% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MQMF.r
\docType{data}
\name{npf}
\alias{npf}
\title{npf fishery catch data from Northern Prawn Fishery 1970-1992}
\format{A data.frame of fisheries data
\describe{
  \item{year}{the fishing year from 1970 - 1992.}
  \item{banana}{banana prawn catches, tonnes.}
  \item{tiger}{tiger prawn catches, tonnes.}
  \item{endevaour}{endeavour prawn catches, tonnes.}
  \item{king}{king prawn catches, tonnes.}
  \item{boats}{the number of vessesl fishing in that year.}
  \item{boatday}{the total annual effort as boatdays.}
}}
\usage{
npf
}
\description{
npf is fishery catch data from Australia's Northern 
    Prawn Fishery from 1970 to 1992 summarized from Robins and
    Somers, 1994. It contains the catches, in 
    tonnes, of banana prawns (Penaeus merguiensis and P. indicus), 
    tiger prawns (brown - P. esculentus) and (grooved - P. 
    semisulcatus), endeavour prawns, (Metapenaeus endevaouri and M.
    ensis), king prawns (P. latisulcatus and P. longistylus), the
    number of vessesl fishing, and the annual effort as boatdays.
}
\examples{
 \dontrun{
 data(npf)
 npf
}
}
\references{
Robins, C. and I. Somers (1994) Appendix A. Fishery 
    Statistics pp 141 - 164 in Pownall, P.C. (ed.) Australia's 
    Northern Prawn Fishery: The first 25 years. NPF25. Cleveland, 
    Austrlaia. 179p.
}
\keyword{datasets}
