% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mqmf_utils.r
\name{parset}
\alias{parset}
\title{parset alters the current base graphics par settings}
\usage{
parset(cex = 0.85, font = 7)
}
\arguments{
\item{cex}{the size of the font used, defaults to 0.85}

\item{font}{the font used, defaults to 7 which is Times Bold, 6 is
Times, 1 is Sans and 2 is Sans Bold.}
}
\value{
nothing but it changes teh base graphics par settings
}
\description{
parset alters the current base graphics par settings
    to suit a single standard plot. It is merely here to simplify
    and speed the coding for exploratory base graphics. The font
    and its size default to 0.85 and font 7 (Times bold). The
    default values can be seen by typing parset with no brackets in
    the console. If a different
    set of par values are needed then the function parsyn() can be
    used to act as a prompt for the correct syntax. The output to
    the console can be copied to your script and modified to suit.
}
\examples{
\dontrun{
parset()
parsyn()
}
}
