% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfuns.R
\name{parsyn}
\alias{parsyn}
\title{parsyn types the standard par command syntax  to the console}
\usage{
parsyn()
}
\value{
it writes two lines of R code to the console
}
\description{
parsyn types the standard par command syntax to the console
     so it can be copied and pasted into your own code and modified.
}
\examples{
\dontrun{
 parsyn()
}
}
