% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfuns.R
\name{plotprep}
\alias{plotprep}
\title{plotprep: sets up a window and the par values for plotting}
\usage{
plotprep(width = 6, height = 3.6, usefont = 7, cex = 0.75,
  newdev = TRUE, filename = "", resol = 300)
}
\arguments{
\item{width}{defaults to 6 inches = 15.24cm - width of plot}

\item{height}{defaults to 3 inches = 7.62cm - height of plot}

\item{usefont}{default=7 (bold Times) 1=sans serif, 2=sans serif bold}

\item{cex}{default=0.75, the size of font used for text within the plots}

\item{newdev}{reuse a previously defined graphics device or make a new 
one; default=TRUE}

\item{filename}{default="" ie do not save to a filename. If a filename 
is defined it creates that file as a png file with resolution resol}

\item{resol}{resolution of the png file if one is defined, default=300}
}
\value{
sets up a graphics device, if needed and sets the default 
    plotting par values. This changes the current plotting options!
}
\description{
plotprep: sets up a window and the par values for plots.
   This is simply a utility function to save typing the standard syntax.
   Some of the defaults can be changed. Typing the name without () will
   provide a template for modification. If different par values are 
   wanted then just include a par statement after plotprep()
}
\examples{
\dontrun{
 x <- rnorm(1000,mean=0,sd=1.0)
 plotprep()
 hist(x,breaks=30,main="",col=2)
 plotprep(width=6,height=5)
 par(mfrow = c(2,1))
 hist(x,breaks=20,main="",col=2)
 hist(x,breaks=30,main="",col=3)
}
}
