% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mqmf_utils.r
\name{properties}
\alias{properties}
\title{properties - used to check a data.frame before standardization}
\usage{
properties(indat, dimout = FALSE)
}
\arguments{
\item{indat}{the data.frame containing the data fields to be used
in the subsequent standardization. It tabulates the number of
NAs and the number of unique values for each variable and finds
the minimum and maximum of the numeric variables}

\item{dimout}{determines whether or noth the dimensions of the data.frame
are printed to the screen or not; defaults to FALSE}
}
\value{
a data.frame with the rows being each variable from the input
    input data.frame and the columns being the number of NAs, the
    number of unique values, and minimum and maximum (where possible).
}
\description{
properties used to check a data.frame before
    standardization
}
\examples{
\dontrun{
data(abdat)
properties(abdat)
}
}
