% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mqmf_utils.r
\name{quants}
\alias{quants}
\title{quants used in apply to estimate quantiles across a vector}
\usage{
quants(invect, probs = c(0.025, 0.05, 0.5, 0.95, 0.975))
}
\arguments{
\item{invect}{vector of values}

\item{probs}{a vector of quantile, default=c(0.025,0.05,0.5,0.95,0.975)}
}
\value{
a vector of the probs quantiles
}
\description{
quants used in 'apply' to estimate quantiles across a vector.
    Can also be used within tapply, etc.
}
\examples{
x <- matrix(rnorm(1000),ncol=10,nrow=100)
apply(x,2,quants)
apply(x,2,quants,probs=c(0.1,0.5,0.9))
}
