% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MQMF.r
\docType{data}
\name{schaef}
\alias{schaef}
\title{schaef is the yellowfin tuna fishery data from Schaefer 1957}
\format{A matrix of fisheries data
\describe{
  \item{year}{the fishing year from 1934 - 1955}
  \item{catch}{the total annual catch, tonnes }
  \item{effort}{ the total effort as standard fishing days}
  \item{cpue}{the catch '000 pounds per standard day, a ratio cpue}
}}
\usage{
schaef
}
\description{
schaef is the yellowfin tuna fishery data from Schaefer, 
    M.B. (1957) A study of the dynamics of the fishery for yellowfin 
    tuna in the Eastern Tropical Pacific Ocean. _Bulletin, Inter-American
    Tropical Tuna Commission_ __2__: 247-285. It contains the year,
    the catch, the effort, and the cpue and was used in one of the first 
    descriptions of a stock assessment that used a surplus production 
    model. The catch-per-unit-effort, cpue, is a ratio cpue of the total
    catch divided by the total effort as thousands of punds per day. 
    These days such ratios tend not to be used with individual records 
    for each day's effort being used instead. 
    This does not obscure the variation between different 
    vessels, areas, depths, and seasons. Initial parameter estimates 
    close to the optimum values for both the Schaefer model and the
    Fox model could be
    param <- log(c(r=0.24,K=2100000,Binit=2200000,sigma=0.2))
}
\examples{
 \dontrun{
 data(schaef)
 schaef
}
}
\keyword{datasets}
