% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spm.r
\name{summspm}
\alias{summspm}
\title{summspm extracts critical statistics from output of displayModel}
\usage{
summspm(ans)
}
\arguments{
\item{ans}{the object output from the function displayModel used to plot
the output of a surplus production analysis}
}
\value{
a matrix of statistics relating to MSY, expected yields, and
    depletion levels
}
\description{
summspm extracts critical statistics from output of displayModel.
    In particular it pulls out the catchability q, teh MSY, Bmsy, Dmsy, Blim,
    Btarg, Ctarg. and Dcurr.
}
\examples{
data(dataspm)
fish <- dataspm$fish
glb <- dataspm$glb
plotfishM(fish,glb)
pars <- c(0.264,4740,3064)
ans <- displayModel(pars,fish,schaefer=FALSE)
bestSP <- optim(par=pars,fn=negLLM,callfun=simpspm,indat=fish,schaefer=FALSE)
outoptim(bestSP)
ans <- displayModel(bestSP$par,fish,schaefer=FALSE)
summspm(ans)
}
